Collapsinator_V2.py
This is a modification of Collapsinator, which ignores the first 12 bases of read 2, and searches for only one spacer. It is useful when reads have low quality in the first few bases of the second read  (which we sometimes observe). It should be used with the following command line flags  :collapsinator_VN2.py -in file -N -mq 0 -aq 0 -bd